refactor(leaderboard): Extract top three cards into a separate component
